src/back-end/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
click==8.1.7
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
pydantic==2.6.2
pydantic_core==2.16.3
pytz==2024.1
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
